time_tracker/requirements.txt
aiogram
python-dotenv
sqlalchemy
streamlit
pandas
plotly
faker